gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = ""
authors = ["Andrew Li <andrewli@u.northwestern.edu>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
polars = "^1.2.1"
bcrypt = "^4.2.0"
pyjwt = "^2.8.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.1"
flask-cors = "^4.0.1"
waitress = "^3.0.0"


[tool.poetry.group.dev.dependencies]
pylint = "^3.2.6"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
